Code dump of AppNexus Common Framework so far + unpackaged common/ code
This C library contains core components on which other higher level
libraries and applications can build.

We have packaged some of our "common" code used by applications in the
real-time platform; that is the original basis for ACF. However,
that's only a small portion of the fundamental code we use in
production, and it's not clear that the current packaging approach is
the best way for us to reuse code while allowing users to decide how
much they would like to opt in our internal framework.

We thus decided to publish most of the remaining shared code under
common/, despite the lack of packaging. Suboptimal packaging should
not stand in the way of progress, and it should be easy for others to
copy from common/ and remove AppNexus-specific noise.  Eventually, all
that should be nicely packaged in the ACF library. In the meantime,
we'd rather share as much as we easily can, even if the source has
depends on unpublished, more specialised, support code.

If we wait until this is perfect, it'll never go out.